GitOps/pyproject.toml
[tool.poetry]
name = "gitops"
version = "0.1.0"
description = "Kluctl"
authors = ["Nicolas signed-log FORMICHELLA <stigpro@outlook.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kubernetes = "^28.0.0"
ruamel-yaml = "^0.18.6"
fabric = "^3.2.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
